added code commenting
