+++ features/comment/index.js
@@ -1,0 +1,6 @@
+const commentor = require('./comment.js');
+
+// Combine and export all functions
+module.exports = {
+  ...commentor
+};
